ragcore-vectordb-server/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
langchain
langchain_postgres
langchain_openai
pymupdf
